requirements.txt
streamlit
google-generativeai
python-dotenv
deep-translator
geopy
haversine
streamlit-audiorecorder
